Giant code push, and reformatting. Forgot to commit, whoopsie

+++ src/main/java/com/wexalian/jtrakt/endpoint/sync/TraktSyncUpdate.java
@@ -4,75 +4,62 @@ import com.wexalian.jtrakt.endpoint.TraktIds;
 
 import java.util.List;
 
-public class TraktSyncUpdate
-{
+public class TraktSyncUpdate {
     private Data added;
     private Data updated;
     private Data existing;
     private Data deleted;
     private NotFound not_found;
     
-    public Data getAdded()
-    {
+    public Data getAdded() {
         return added;
     }
     
-    public Data getUpdated()
-    {
+    public Data getUpdated() {
         return updated;
     }
     
-    public Data getExisting()
-    {
+    public Data getExisting() {
         return existing;
     }
     
-    public Data getDeleted()
-    {
+    public Data getDeleted() {
         return deleted;
     }
     
-    public NotFound getNotFound()
-    {
+    public NotFound getNotFound() {
         return not_found;
     }
     
-    public static class Data
-    {
+    public static class Data {
         private int movies;
         private int shows;
         private int seasons;
         private int episodes;
         private int people;
         
-        public int getMovies()
-        {
+        public int getMovies() {
             return movies;
         }
         
-        public int getShows()
-        {
+        public int getShows() {
             return shows;
         }
         
-        public int getSeasons()
-        {
+        public int getSeasons() {
             return seasons;
         }
         
-        public int getEpisodes()
-        {
+        public int getEpisodes() {
             return episodes;
         }
         
-        public int getPeople()
-        {
+        public int getPeople() {
             return people;
         }
     }
     
-    public static class NotFound
-    {
+    public static class NotFound {
         private List<Item> movies;
         private List<Item> shows;
         private List<Item> seasons;
@@ -80,48 +67,39 @@ public class TraktSyncUpdate
         private List<Item> people;
         private List<Long> ids;
         
-        public List<Item> getMovies()
-        {
+        public List<Item> getMovies() {
             return movies;
         }
         
-        public List<Item> getShows()
-        {
+        public List<Item> getShows() {
             return shows;
         }
         
-        public List<Item> getSeasons()
-        {
+        public List<Item> getSeasons() {
             return seasons;
         }
         
-        public List<Item> getEpisodes()
-        {
+        public List<Item> getEpisodes() {
             return episodes;
         }
         
-        public List<Item> getPeople()
-        {
+        public List<Item> getPeople() {
             return people;
         }
         
-        public List<Long> getIds()
-        {
+        public List<Long> getIds() {
             return ids;
         }
         
-        public static class Item
-        {
+        public static class Item {
             private TraktIds ids;
-            private int rating; //only used for ratings
+            private float rating; //only used for ratings
             
-            public TraktIds getIds()
-            {
+            public TraktIds getIds() {
                 return ids;
             }
             
-            public int getRating()
-            {
+            public float getRating() {
                 return rating;
             }
         }
